Change implementation to a ksp module instead of a feature toggle
